# wpfclient/web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "web, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Items": {
      "get": {
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToDoItems"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToDoItems"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToDoItems"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Items/{id}/getitem": {
      "get": {
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoItems"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoItems"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoItems"
                }
              }
            }
          }
        }
      }
    },
    "/Items/AddItems": {
      "post": {
        "tags": [
          "Items"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToDoItems"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ToDoItems"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ToDoItems"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Items/deleteitem/{id}": {
      "delete": {
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ToDoItems": {
        "type": "object",
        "properties": {
          "toDoItemsId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "dueTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}